# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is balance service for transfer money between user accounts.",
        "title": "Balance Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/balance": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Get user account balance",
                "operationId": "balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Currency code",
                        "name": "currency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user balance",
                        "schema": {
                            "$ref": "#/definitions/mapper.Balance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transfer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Transfer money between user accounts",
                "operationId": "transfer",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mapper.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transfers": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Get list transfers for filter",
                "operationId": "transfers",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "amountEQ",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "amountGTE",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Amount filter",
                        "name": "amountLTE",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date filter",
                        "name": "dateFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "dateTo",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "transferType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "userID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list transfers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mapper.Transfer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "mapper.Balance": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "money": {
                    "$ref": "#/definitions/mapper.Money"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "mapper.Money": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "code": {
                    "description": "ISO 4217",
                    "type": "string"
                }
            }
        },
        "mapper.Transfer": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "mapper.TransferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "receiver_id": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                }
            }
        }
    }
}